rename validate_header module
