+++ crates/amaru-consensus/src/consensus/validate_header.rs
